organization changes before reducing imports
