+++ src/rust/graph-generator-lib/src/serialization.rs
@@ -1,0 +1,74 @@
+use graph_descriptions::graph_description::*;
+use log::*;
+use sqs_lambda::completion_event_serializer::CompletionEventSerializer;
+use std::io::Cursor;
+
+/// Takes a subgraph generated by a Generator and serializes it so it can be returned by a
+/// lambda as an execution result.
+#[derive(Clone, Debug, Default)]
+pub struct SubgraphSerializer {
+    proto: Vec<u8>,
+}
+
+impl SubgraphSerializer {
+    pub fn new(proto: Vec<u8>) -> Self {
+        Self { proto }
+    }
+}
+
+impl CompletionEventSerializer for SubgraphSerializer {
+    type CompletedEvent = Graph;
+    type Output = Vec<u8>;
+    type Error = sqs_lambda::error::Error;
+
+    fn serialize_completed_events(
+        &mut self,
+        completed_events: &[Self::CompletedEvent],
+    ) -> Result<Vec<Self::Output>, Self::Error> {
+        let mut subgraph = Graph::new(0);
+
+        let mut pre_nodes = 0;
+        let mut pre_edges = 0;
+
+        for sg in completed_events.iter() {
+            pre_nodes += sg.nodes.len();
+            pre_edges += sg.edges.len();
+            subgraph.merge(sg);
+        }
+
+        if subgraph.is_empty() {
+            warn!(
+                concat!(
+                    "Output subgraph is empty. Serializing to empty vector.",
+                    "pre_nodes: {} pre_edges: {}"
+                ),
+                pre_nodes, pre_edges,
+            );
+            return Ok(vec![]);
+        }
+
+        info!(
+            "Serializing {} nodes {} edges. Down from {} nodes {} edges.",
+            subgraph.nodes.len(),
+            subgraph.edges.len(),
+            pre_nodes,
+            pre_edges,
+        );
+
+        let subgraphs = GeneratedSubgraphs {
+            subgraphs: vec![subgraph],
+        };
+
+        self.proto.clear();
+
+        prost::Message::encode(&subgraphs, &mut self.proto)
+            .map_err(|e| sqs_lambda::error::Error::EncodeError(e.to_string()))?;
+
+        let mut compressed = Vec::with_capacity(self.proto.len());
+        let mut proto = Cursor::new(&self.proto);
+        zstd::stream::copy_encode(&mut proto, &mut compressed, 4)
+            .map_err(|e| sqs_lambda::error::Error::EncodeError(e.to_string()))?;
+
+        Ok(vec![compressed])
+    }
+}
